Move VersionParser tests out of VersionRequirementTest
