+++ Assets/Plugins/Editor/Uplift/Testing/UnitTesting/VersionTest.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: a15d2129f236adc4e8ea2f982a78df93
+timeCreated: 1512727612
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
